@@ -3,7 +3,7 @@
     "description": "Simple PHP abstraction of template rendering.",
     "type": "library",
     "require": {
-        "php": "^5.6 || ^7.0"
+        "php": "^7.4 || ^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -26,42 +26,7 @@
     },
     "minimum-stability": "stable",
     "require-dev": {
-        "phpunit/phpunit": "~5.0",
-        "friendsofphp/php-cs-fixer": "^2.14"
-    },
-    "scripts": {
-        "check-cs": [
-            "vendor/bin/php-cs-fixer fix -vvv --dry-run --ansi --config php-cs-fixer/config.php src",
-            "vendor/bin/php-cs-fixer fix -vvv --dry-run --ansi --config php-cs-fixer/config.php tests"
-        ],
-        "check-staged-cs": [
-            "vendor/bin/php-cs-fixer fix -vvv --dry-run --ansi --config php-cs-fixer/config.php"
-        ],
-        "diff-cs": [
-            "vendor/bin/php-cs-fixer fix -vvv --dry-run --diff --ansi --config php-cs-fixer/config.php src",
-            "vendor/bin/php-cs-fixer fix -vvv --dry-run --diff --ansi --config php-cs-fixer/config.php tests"
-        ],
-        "diff-staged-cs": [
-            "vendor/bin/php-cs-fixer fix -vvv --dry-run --diff --ansi --config php-cs-fixer/config.php"
-        ],
-        "fix-cs": [
-            "vendor/bin/php-cs-fixer fix -vvv --ansi --config php-cs-fixer/config.php src",
-            "vendor/bin/php-cs-fixer fix -vvv --ansi --config php-cs-fixer/config.php tests"
-        ],
-        "fix-json": [
-            "mv composer.json composer.json.bck",
-            "cat composer.json.bck | python -mjson.tool > composer.json",
-            "diff composer.json.bck composer.json -u && rm composer.json.bck"
-        ],
-        "fix-staged-cs": [
-            "vendor/bin/php-cs-fixer fix -vvv --ansi --config php-cs-fixer/config.php"
-        ],
-        "pre-commit": [
-            "composer check-staged-cs",
-            "composer test"
-        ],
-        "test": [
-            "vendor/bin/phpunit --colors=always"
-        ]
+        "phpunit/phpunit": "^9.1",
+        "squizlabs/php_codesniffer": "^3.5"
     }
 }
